Rip out the Infinite movie feature, it was a gimmick that has run its course. It isn't much code but it does complicate the API for a feature that would never lead to a published tas movie

+++ src/BizHawk.Client.Common/Api/Classes/MovieApi.cs
@@ -98,7 +98,7 @@ namespace BizHawk.Client.Common
 
 		public bool IsLoaded() => Global.MovieSession.Movie.IsActive();
 
-		public double Length() => Global.MovieSession.Movie.FrameCount;
+		public int Length() => Global.MovieSession.Movie.FrameCount;
 
 		public string Mode() => Global.MovieSession.Movie.Mode.ToString().ToUpper();
